requirements.txt
django<6
python-dotenv==1.0.1
mysqlclient==2.2.5
